feat: static sitemap generation during build
