--- apps/daffio/src/app/core/sitemap/index.ts
@@ -1,3 +1,0 @@
-export { DaffioSitemapService } from './sitemap.service';
-export { DaffioDocsSitemapService } from './docs-sitemap.service';
-export type { SitemapUrl } from './sitemap.service';
